Keymanagement: Handle MariaDB constraint limitation.
MariaDB has some limitations in how "unique together" is managed.
It's not possible to have a migration containing four fields, at
least not if one of those are also a foreign key. It is possible
to add all four desired fields, if it's done in two or more
migrations, but MariaDB will fail to enforce the constraint.

Workaround the issue by excluding the user field. SSH keys are
already required to be unique globally, eliminating the possibility
that any two users have the same SSH key.

Move handling for deactivating other keys for the same system to a
signal, ensuring that only one key can be active for one system
simultaneously, even when set via the Django admin interface, and
in the case MariaDB fails to handle the constraint violation.

Change-Id: I9630f57142c85389006a4a4fe6635b2669a53d3f