Keymanagement: SSH keys are in some cases not synced to LDAP.
Rework the LDAP interaction to happen in a single task when updating SSH
keys. This avoids conditions where a key is automatically disabled in
Bitu before it can be syncronised to LDAP. By handling the entire
syncronization process in a single tasks we ensure that all operations
happen in the correct order.

Bug: T351139
Change-Id: I6a4cfa35f8e8c7ffc755a683134b796673f89400