Update error pages to Codex design.
* Remove old Bulma files.
* Update custom error pages to Codex

This patch ensures that we will be able to run Django REST framework
tests in the future. The test API client will parse the custom error
pages, but fail to get the static files used by the Bulma styled pages,
because those files are not included in the static collection config.

Change-Id: Ife5519ca20f95ea83748f1dce7a8906ce9573b1b