Build Bitu contain image using Blubber.
Provide the option of running Bitu as a Docker image.
Current iteration make a number of assumptions about the state of the
LDAP directory in order to facilitate easier usage.

Users can, if required mount their own settings.py when starting the
container.

Initial LDAP data have been updated to include an admin user, with the
default password of "admin".

Docker-compose.yaml is updated to automatically load a full environment,
but can still be used for development purposes.

Bug: T362318
Change-Id: I0be68d0464503207ed824380f77fbebfc0fa6b3a