Replace development server with uWSGI.
The container image is going to be used to run instances of Bitu, in
those cases using the development server would not be recommended as it,
amount other issues, is single threated.

For development work it cannot be assumed that the user will have a
webserver in front, proxying the uWSGI trafic. In those cases allow
uWSGI to start up in embedded http mode.

- HTTP mode is enabled by setting UWSGI_HTTP_MODE=true
- UWSGI process count is adjusted with UWSGI_PROCESS_COUNT, default is 4
- UWSGI/HTTP port can be changed using UWSGI_PORT, default is 8080
  This should not be needed as the port exposed port can be remapped by
  Docker

Change-Id: I000f3f97d2c245eeaa018cb4e4108cebac69fe9c