MediaWiki: Allow Bitu to be used as a 2FA proxy.
Given an LDAP username, Bitu will lookup the users SUL account name
and perform a validation of a provided TOTP token, using the MediaWiki
oathauth extension.

Bitu will return the provided username, the SUL account name, if the
user have 2FA enabled and if the provided token is valid or not. If the
user does not have a linked SUL account the field is left blank and
validation of the TOTP token is skipped.

The rest-endpoint is available as /mediawiki/api/totp and protected
using Django permissions.

To enable the API, Bitu must have ENABLE_API set to True and the
MediaWiki integration configured as:

MEDIAWIKI = {
    'host': 'https://mediawiki.example.org',
    'consumer_token': 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
    'consumer_secret': 'yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy',
    'access_token': 'vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv',
    'access_secret': 'zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz',
    'user_agent': 'Bitu IDM',
}

where the consumer and access tokens/secrets are a OAuth1a client
configured in MediaWiki.

Bug: T359551
Change-Id: Ic0be92f182483779e19257371baef75ce991fb67