Show currently signed in username.
Allow the users to easily see which username they are signed in as.
Users can already see their shell account username, but not their CN
from LDAP/CAS. Show the username the user used to signin in the
navigation menu and add it to the listed attributes in the developer
environment.

Bug: T378344
Change-Id: I0290acd65b06969220fcbfa62acedaeb56d4fa09