Prevalidation of permissions
For some type of permissions it makes sense to do the validation of the
user upfront, and only show the permission as "request-able" if there is
a chance that the request will be approved.

E.g. it does not make sense to present a permission that requires a
certain LDAP permission or email domain, if the user is known to not be
able to fullfill that requirement.

Change-Id: I6e27c0f39fdcf7060c4e3b2c8ab9aa535cc77513