Password update: avoid triggering invalid hash error
The passlib module is used for verifying LDAP passwords when user want
to update their password. The verify method assumes that the hash stored
in LDAP will be prefixed with {SSHA} (for SHA1) hashes. Other libraries
and services may opt to store the hash with {ssha} has the prefix.
Passlib will see this as an invalid hash and fail to accept the old
password and trigger a http 500 in Bitu.

* Check for the lower case ssha and replace if require.
* Handle any cases where a hash may be invalid for other reasons and
  inform the user.
* Add logging to help identify future issues with the validator.
* Add test cases for the new code.

Bug: T381327
Change-Id: Icf1febc638da65002b727b3ef7e92aa0230afbce