Updated notification handling
Separate notification from managers. This allows us to define a set of
managers who can approve a permission request, while having a separate
set of managers, groups or individuals being notified of the request.

In most cases the notify_group will be identical to the LDAP group that
can approve the requests.

Bug: T381075
Change-Id: I5d5b407d0842bf44679f2a37517cdabe2e28f5be