Password reset: use passlib for hashing
The Bitu LDAP library uses a low level function for hashing passwords,
this results in the hash being set as ssha, rather than SSHA. Most
libraries will understand that these are the same, but passlib does not.

To ensure that Bitu does not generate password hashes it can not
validate later, switch to also use passlib for the hashing when users
reset their password.

Bug: T381327
Change-Id: Icfcebe965eceee61e3304a040c4b9e2f34ca82ed