Disallow emails as username
Implement validator to prevent users from signing up using their email
address as a username. There's nothing in LDAP that prevent or
discourage the use of emails as CN, but in some circumstances it can
lead to confusing.

E.g. we do not allow usernames to change, meaning that a user may be
stuck with an account using their old email address as a username, but a
completely different address as their point of contact.

In addition, the account signup system used before idm.wikimedia.org
(implemented via a MediaWiki extension running on wikitech.wikimedia.org)
enforced that usernames were following Mediawiki username syntax,
also prohibiting the use of @ signs.

Bug: T382226
Change-Id: I529391d684adaf6df29de480809e5bd14a6afb65