Update unittests to handle BituLDAP update
The next release of the BituLDAP library will always create a new
connection, rather than attempting to reuse a shared connection. This is
done to avoid issue with reconnecting.

This change means that we're no longer able to simply inject a
mock connection objecting into the library, and instead the unittests must
patch out the create_connection method.

Change-Id: Iabd9168f320985cc4e7f3033f3a3ab08a6d1e0db