Adjust build.sh for other environments
On a Linux x86_64 I encountered a build failure:

    .buildkit_qemu_emulator: /bin/sh:
    Invalid ELF image for this architecture

Pass to `--platform` the machine name from `uname -m` though I would
expect Docker to do it by default?

The `testing` Blubber variant has been renamed `test`.

Prefix the image tag with `localhost/` to make it clear it is a local
build image and not one that is published or coming from a remote
registry.

Publish the image upon building (`--load`), Docker does it automatically
but that is not the case when using an alternative system such as
Podman.

Change-Id: I0820ee36a8d328b1433315fa09df6a1bfcd7cfb9