Block Gerrit users with uid rather than cn
After I have requested a legacy bot account to be blocked, I have
discovered the integration with Gerrit was not working and this was
addressed by porting the code that was previously in mediawiki-config.

The Bitu integration uses the LDAP `cn` field to block the user which is
incorrect as we as previously investigated on T307558. Instead users
should be blocked using the `uid` field which is guaranteed to be unique
and is not subject to a fuzzy search.

Change Gerrit integration to use `uid` rather than `cn`.

Bug: T390070
Change-Id: I587deaf7b76c68e6522006373b2c761fedc842cc