Signup: command for generating activation link
In some cases it may not be possible for a user to receive an activation
link for their account creation. This can happen in their email provide
decides to discard email from us. In those cases all operators with
shell access to generate an activation link, which can then be
distributed via other channels.

Change-Id: I64199220a9adf02606e6b12723da6791ea4cd06e